moved Event.java to logic and made SimpleDateFormat static field in Event.java public
