+++ app/src/main/java/ch/epfl/polycrowd/EventPageActivity.java
@@ -8,6 +8,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 import ch.epfl.polycrowd.firebase.FirebaseInterface;
+import ch.epfl.polycrowd.logic.Event;
 
 import android.os.Bundle;
 import android.view.Menu;
@@ -15,8 +16,6 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.widget.SearchView;
 import android.widget.Toast;
-
-import java.text.ParseException;
 
 public class EventPageActivity extends AppCompatActivity {
