+++ app/src/main/java/ch/epfl/polycrowd/OrganizerInviteActivity.java
@@ -9,6 +9,8 @@ import android.widget.TextView;
 
 import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
+
+import ch.epfl.polycrowd.logic.Event;
 import ch.epfl.polycrowd.logic.PolyContext;
 
 public class OrganizerInviteActivity extends AppCompatActivity {
