+++ app/src/main/java/ch/epfl/polycrowd/firebase/handlers/EventHandler.java
@@ -1,6 +1,6 @@
 package ch.epfl.polycrowd.firebase.handlers;
 
-import ch.epfl.polycrowd.Event;
+import ch.epfl.polycrowd.logic.Event;
 
 public interface EventHandler {
     void handle(Event event);
