+++ app/src/main/java/ch/epfl/polycrowd/firebase/handlers/EventsHandler.java
@@ -2,7 +2,7 @@ package ch.epfl.polycrowd.firebase.handlers;
 
 import java.util.List;
 
-import ch.epfl.polycrowd.Event;
+import ch.epfl.polycrowd.logic.Event;
 
 public interface EventsHandler {
     void handle(List<Event> events);
